Started on refactor of zonesearchbymapscreen
